move ftype
