+++ src/finchlite/codegen/c_codegen.py
@@ -17,7 +17,7 @@ from finchlite.algebra import ffunc
 from finchlite.algebra.algebra import FinchOperator
 
 from .. import finch_assembly as asm
-from ..algebra import COperator, query_property, register_property
+from ..algebra import COperator, FType, fisinstance, ftype, query_property, register_property
 from ..finch_assembly import (
     AssemblyStructFType,
     BufferFType,
@@ -26,7 +26,7 @@ from ..finch_assembly import (
     MutableStructFType,
     TupleFType,
 )
-from ..symbolic import Context, FType, Namespace, ScopedDict, fisinstance, ftype
+from ..symbolic import Context, Namespace, ScopedDict
 from ..util import config, file_cache
 from ..util.logging import LOG_BACKEND_C
 from .stages import CCode, CLowerer
