+++ src/finchlite/finch_logic/interpreter.py
@@ -7,8 +7,7 @@ import finchlite
 from finchlite.algebra.tensor import TensorFType
 from finchlite.finch_assembly import AssemblyKernel, AssemblyLibrary
 
-from ..algebra import fixpoint_type, return_type
-from ..symbolic import fisinstance
+from ..algebra import fisinstance, fixpoint_type, return_type
 from ..util.logging import LOG_LOGIC_PRE_OPT
 from . import nodes as lgc
 from .nodes import (
