+++ tests/test_assembly_type_checker.py
@@ -6,11 +6,11 @@ import numpy as np
 
 import finchlite.finch_assembly as asm
 from finchlite import ffunc
+from finchlite.algebra import FType, ftype
 from finchlite.codegen import NumpyBuffer
 from finchlite.codegen.hashtable import CHashTable, NumbaHashTable
 from finchlite.finch_assembly import assembly_check_types
 from finchlite.finch_assembly.struct import TupleFType
-from finchlite.symbolic import FType, ftype
 
 
 def test_lit_basic():
